# model-service/openapi.json
{
  "components": {
    "schemas": {
...
      "AugmentRequest": {
        "description": "Request body for the ontology augmentation endpoint.",
        "properties": {
          "domain": {
            "description": "Domain description for context",
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "existing_types": {
            "description": "Existing type names",
            "items": {
              "type": "string"
            },
            "title": "Existing Types",
            "type": "array"
          },
          "max_suggestions": {
            "default": 10,
            "description": "Maximum suggestions to return",
            "maximum": 50,
            "minimum": 1,
            "title": "Max Suggestions",
            "type": "integer"
          },
          "persona_id": {
            "description": "Unique identifier for the persona",
            "minLength": 1,
            "title": "Persona Id",
            "type": "string"
          },
          "target_category": {
            "description": "Category to augment",
            "enum": [
              "entity",
              "event",
              "role",
              "relation"
            ],
            "title": "Target Category",
            "type": "string"
          }
        },
        "required": [
          "persona_id",
          "domain",
          "target_category"
        ],
        "title": "AugmentRequest",
        "type": "object"
      },
      "AugmentResponse": {
        "description": "Response body for the ontology augmentation endpoint.",
        "properties": {
          "id": {
            "description": "Unique identifier for this augmentation",
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "persona_id": {
            "description": "Persona identifier",
            "minLength": 1,
            "title": "Persona Id",
            "type": "string"
          },
          "reasoning": {
            "description": "Explanation of why these types were suggested",
            "title": "Reasoning",
            "type": "string"
          },
          "suggestions": {
            "description": "Suggested types",
            "items": {
              "$ref": "#/components/schemas/OntologyType"
            },
            "title": "Suggestions",
            "type": "array"
          },
          "target_category": {
            "description": "Category that was augmented",
            "title": "Target Category",
            "type": "string"
          }
        },
        "required": [
          "id",
          "persona_id",
          "target_category",
          "reasoning"
        ],
        "title": "AugmentResponse",
        "type": "object"
      },
...
      "OntologyType": {
        "description": "Suggested ontology type from augmentation.",
        "properties": {
          "confidence": {
            "default": 0.0,
            "description": "Confidence score",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "description": "Type description",
            "title": "Description",
            "type": "string"
          },
          "examples": {
            "description": "Example instances",
            "items": {
              "type": "string"
            },
            "title": "Examples",
            "type": "array"
          },
          "name": {
            "description": "Type name",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Parent type name",
            "title": "Parent"
          },
          "thinking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThinkingTrace"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional reasoning trace from a thinking-capable model"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "OntologyType",
        "type": "object"
      },
...
      "ThinkingStep": {
        "description": "One step of a chain-of-thought trace.",
        "properties": {
          "content": {
            "description": "Reasoning step text",
            "title": "Content",
            "type": "string"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional token count",
            "title": "Tokens Used"
          }
        },
        "required": [
          "content"
        ],
        "title": "ThinkingStep",
        "type": "object"
      },
      "ThinkingTrace": {
        "description": "Captured reasoning trace from a thinking-capable model.",
        "properties": {
          "model_id": {
            "default": "",
            "description": "Producing model identifier",
            "title": "Model Id",
            "type": "string"
          },
          "steps": {
            "description": "Reasoning steps in order",
            "items": {
              "$ref": "#/components/schemas/ThinkingStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional total tokens across steps",
            "title": "Total Tokens"
          }
        },
        "title": "ThinkingTrace",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Cross-service contract for the request/response shapes the FOVEA model-service produces and the server consumes. Generated without importing the ML stack by scripts/gen_contract_spec.py; do not edit by hand.",
    "title": "FOVEA Model Service Contract",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ontology/augment": {
      "post": {
        "operationId": "post_/api/ontology/augment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AugmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AugmentResponse"
                }
              }
            },
            "description": "Successful response."
          }
        },
        "summary": "Suggest ontology types for a persona domain."
      }
    },
...
  }
}